app/requirements.txt
flask
Flask-BasicAuth
gunicorn
python-dateutil
requests